test(factory-sessions): preserve split file budget
