+++ pkg/services/factory_sessions/internal/execution/fake_service_construction_start_test.go
@@ -4,13 +4,19 @@ import (
 	"context"
 	"encoding/json"
 	"errors"
-	"os"
-	"strings"
-	"sync"
-	"testing"
-
+	"fmt"
+	"github.com/portpowered/infinite-you/internal/testutil/checkpointfixtures"
 	"github.com/portpowered/infinite-you/pkg/services/factory_runtime"
 	"github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/fileeffects"
+	"github.com/portpowered/infinite-you/pkg/services/workers"
+	"os"
+	"path/filepath"
+	"reflect"
+	"strings"
+	"sync"
+	"sync/atomic"
+	"testing"
+	"time"
 )
 
 func TestNewFakeServiceRequiresExplicitClockBeforeFixtureIO(t *testing.T) {
@@ -554,4 +560,299 @@ func TestFakeService_ConstructorsAndHelpers(t *testing.T) {
 	if _, ok := findDispatchSummary([]DispatchSummary{{ID: "disp-1"}}, "missing"); ok {
 		t.Fatal("findDispatchSummary should report missing dispatch")
 	}
+}
+
+var executionTestSessionIdentity atomic.Uint64
+
+func testSessionIDGenerator() string {
+	return fmt.Sprintf("00000000-0000-4000-8000-%012d", executionTestSessionIdentity.Add(1))
+}
+
+// constructorWorkflowContracts proves that constructor validation receives the
+// three Factory Runtime root roles. These tests do not execute their methods.
+type constructorWorkflowContracts struct {
+	factory.JavaScriptWorkflowDefinitions
+	factory.JavaScriptWorkflowRuntime
+	factory.JavaScriptChildValues
+}
+
+func (c constructorWorkflowContracts) RunJavaScript(
+	ctx context.Context,
+	req factory.JavaScriptRuntimeRequest,
+	hooks factory.JavaScriptRuntimeHooks,
+) (factory.JavaScriptRuntimeOutcome, error) {
+	return c.Run(ctx, req, hooks)
+}
+
+func (c constructorWorkflowContracts) ResumeJavaScript(
+	summary factory.JavaScriptCompletedCheckpointSummary,
+	records []factory.JavaScriptRuntimeRecord,
+) factory.JavaScriptResumeContext {
+	return c.ResumeContext(summary, records)
+}
+
+// serviceConfig keeps table-driven constructor tests compact without restoring
+// a production dependency bag.
+type serviceConfig struct {
+	ProjectRoot       string
+	ChildExecutorMode string
+	Provider          workers.Provider
+	ProviderExecutor  workers.InvocationExecutor
+	FakeScenarios     []FakeScenario
+	Persistence       PersistenceChoice
+	Clock             factory.Clock
+	WorkerPresetIDs   map[string]struct{}
+	WorkerSettings    factory.JavaScriptWorkerSettings
+}
+
+func newExecutionService(provider ExecutionProvider, config serviceConfig) (Service, error) {
+	switch provider {
+	case ExecutionProviderFake:
+		clock := config.Clock
+		if clock == nil {
+			clock = durableFixedClock{now: time.Date(2026, 1, 1, 0, 0, 0, 0, time.UTC)}
+		}
+		return NewFakeService(clock, config.FakeScenarios...)
+	case ExecutionProviderJavaScriptRuntime:
+		workflows := constructorWorkflowContracts{}
+		return NewJavaScriptExecutionService(
+			config.ProjectRoot,
+			config.ChildExecutorMode,
+			firstInvocationExecutor(config.ProviderExecutor, config.Provider),
+			config.Persistence,
+			config.Clock,
+			testSyncWaitScheduler{},
+			checkpointfixtures.CheckpointSummariesFixture{
+				BuildResult:  checkpointfixtures.ResumableCheckpointSummaryResult(),
+				LatestResult: checkpointfixtures.ResumableCheckpointSummaryResult(),
+			},
+			workflows,
+			workflows,
+			workflows,
+			config.WorkerPresetIDs,
+			config.WorkerSettings,
+			mustTestRecordingWriter(),
+			testSessionIDGenerator,
+			nil, nil,
+		)
+	default:
+		return nil, NewValidationError("provider", "unsupported execution provider")
+	}
+}
+
+func firstInvocationExecutor(executor workers.InvocationExecutor, provider workers.Provider) workers.InvocationExecutor {
+	if executor != nil {
+		return executor
+	}
+	if provider == nil {
+		return nil
+	}
+	return constructorInvocationExecutor{}
+}
+
+// constructorInvocationExecutor is an inert root-contract value. Constructor
+// tests validate dependency presence only; Workers owns invocation behavior.
+type constructorInvocationExecutor struct {
+	workers.InvocationExecutor
+}
+
+type testSyncWaitScheduler struct{}
+
+func (testSyncWaitScheduler) Now() time.Time { return time.Now() }
+
+func (testSyncWaitScheduler) After(duration time.Duration) <-chan time.Time {
+	return time.After(duration)
+}
+
+func TestNewJavaScriptExecutionServiceRequiresSyncWaitScheduler(t *testing.T) {
+	t.Parallel()
+	workflows := constructorWorkflowContracts{}
+	_, err := NewJavaScriptExecutionService(
+		t.TempDir(),
+		ChildExecutorModeFake,
+		nil,
+		DisabledPersistence(),
+		durableFixedClock{now: time.Date(2026, 1, 1, 0, 0, 0, 0, time.UTC)},
+		nil,
+		checkpointfixtures.CheckpointSummariesFixture{},
+		workflows,
+		workflows,
+		workflows,
+		nil,
+		factory.JavaScriptWorkerSettings{},
+		mustTestRecordingWriter(),
+		testSessionIDGenerator,
+		nil, nil,
+	)
+	if err == nil || !strings.Contains(err.Error(), "sync wait scheduler is required") {
+		t.Fatalf("NewJavaScriptExecutionService error = %v, want missing sync wait scheduler", err)
+	}
+}
+
+func TestWaitSyncCompletionUsesInjectedClockAndRecurringScheduler(t *testing.T) {
+	t.Parallel()
+	clock := newControlledSyncWaitClock(time.Date(2026, 1, 1, 0, 0, 0, 0, time.UTC))
+	service := &JavaScriptRuntimeService{
+		clock:     clock,
+		syncWaits: clock,
+		sessions: map[string]*runtimeSessionState{
+			"session-1": {session: SessionReadResult{SessionID: "session-1", Status: LifecycleStatusRunning}},
+		},
+	}
+
+	result := make(chan SyncStartResult, 1)
+	errs := make(chan error, 1)
+	go func() {
+		got, err := service.waitSyncCompletion(context.Background(), "session-1", 20*time.Millisecond, false)
+		result <- got
+		errs <- err
+	}()
+
+	if duration := <-clock.requests; duration != 10*time.Millisecond {
+		t.Fatalf("first scheduled wait = %s, want 10ms", duration)
+	}
+	clock.Advance(10 * time.Millisecond)
+	if duration := <-clock.requests; duration != 10*time.Millisecond {
+		t.Fatalf("recurring scheduled wait = %s, want 10ms", duration)
+	}
+	select {
+	case got := <-result:
+		t.Fatalf("wait completed before injected deadline: %#v", got)
+	default:
+	}
+
+	clock.Advance(10 * time.Millisecond)
+	if err := <-errs; err != nil {
+		t.Fatalf("waitSyncCompletion: %v", err)
+	}
+	got := <-result
+	if !got.TimedOut || got.SyncOutcome != SyncOutcomeTimedOut {
+		t.Fatalf("result = %#v, want injected-clock timeout", got)
+	}
+}
+
+type controlledSyncWaitClock struct {
+	mu       sync.Mutex
+	now      time.Time
+	waiters  []chan time.Time
+	requests chan time.Duration
+}
+
+func newControlledSyncWaitClock(now time.Time) *controlledSyncWaitClock {
+	return &controlledSyncWaitClock{now: now, requests: make(chan time.Duration, 4)}
+}
+
+func (c *controlledSyncWaitClock) Now() time.Time {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+	return c.now
+}
+
+func (c *controlledSyncWaitClock) After(duration time.Duration) <-chan time.Time {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+	waiter := make(chan time.Time, 1)
+	c.waiters = append(c.waiters, waiter)
+	c.requests <- duration
+	return waiter
+}
+
+func (c *controlledSyncWaitClock) Advance(duration time.Duration) {
+	c.mu.Lock()
+	c.now = c.now.Add(duration)
+	now := c.now
+	waiters := c.waiters
+	c.waiters = nil
+	c.mu.Unlock()
+	for _, waiter := range waiters {
+		waiter <- now
+	}
+}
+
+func contractFixturesPath(t *testing.T) string {
+	t.Helper()
+	return filepath.Join("..", "..", "..", "..", "transports", "http", "testdata", "durable-session-contract-fixtures.json")
+}
+
+func newContractFakeService(t *testing.T) *FakeService {
+	t.Helper()
+	service, err := NewFakeServiceFromContractFixtures(contractFixturesPath(t), fakeServiceTestClock(), fileeffects.ContractFixtureReader(os.ReadFile))
+	if err != nil {
+		t.Fatalf("NewFakeServiceFromContractFixtures: %v", err)
+	}
+	return service
+}
+
+func fakeServiceTestClock() durableFixedClock {
+	return durableFixedClock{now: time.Date(2026, 1, 1, 0, 0, 0, 0, time.UTC)}
+}
+
+func mustNewFakeService(t *testing.T, scenarios ...FakeScenario) *FakeService {
+	t.Helper()
+	service, err := NewFakeService(fakeServiceTestClock(), scenarios...)
+	if err != nil {
+		t.Fatalf("NewFakeService: %v", err)
+	}
+	return service
+}
+
+func int64Ptr(value int64) *int64 {
+	return &value
+}
+
+func assertCanonicalEventEnvelope(t *testing.T, raw json.RawMessage, eventType, id string) {
+	t.Helper()
+	var envelope struct {
+		SchemaVersion string `json:"schemaVersion"`
+		ID            string `json:"id"`
+		Type          string `json:"type"`
+		Context       struct {
+			Sequence int `json:"sequence"`
+		} `json:"context"`
+		Payload json.RawMessage `json:"payload"`
+	}
+	if err := json.Unmarshal(raw, &envelope); err != nil {
+		t.Fatalf("Unmarshal event: %v", err)
+	}
+	if envelope.SchemaVersion != canonicalFactoryEventSchemaVersion {
+		t.Fatalf("schemaVersion = %q, want %q", envelope.SchemaVersion, canonicalFactoryEventSchemaVersion)
+	}
+	if id != "" && envelope.ID != id {
+		t.Fatalf("id = %q, want %q", envelope.ID, id)
+	}
+	if eventType != "" && envelope.Type != eventType {
+		t.Fatalf("type = %q, want %q", envelope.Type, eventType)
+	}
+	if envelope.Context.Sequence <= 0 {
+		t.Fatalf("sequence = %d, want positive", envelope.Context.Sequence)
+	}
+	if len(envelope.Payload) == 0 {
+		t.Fatal("payload missing")
+	}
+}
+
+func jsonEqual(left, right json.RawMessage) bool {
+	var leftValue any
+	var rightValue any
+	return json.Unmarshal(left, &leftValue) == nil &&
+		json.Unmarshal(right, &rightValue) == nil &&
+		reflect.DeepEqual(leftValue, rightValue)
+}
+
+func canonicalTypedInternalEvent(t *testing.T, eventType, sessionID string, payload any) json.RawMessage {
+	t.Helper()
+	raw, err := json.Marshal(map[string]any{
+		"schemaVersion": canonicalFactoryEventSchemaVersion,
+		"id":            "internal/" + eventType,
+		"type":          eventType,
+		"context": map[string]any{
+			"sequence":  99,
+			"sessionId": sessionID,
+		},
+		"payload": payload,
+	})
+	if err != nil {
+		t.Fatalf("marshal %s event: %v", eventType, err)
+	}
+	return raw
 }
